pyproject.toml
[tool.poetry]
name = "showcase-3d-printer-anomaly-detection"
version = "0.1.0"
description = "Anomaly Detection for 3D-Printer Showcase"
license = "Apache-2.0"
authors = [
    "Julian Lengelsen <julian.lengelsen@th-koeln.de>",
    "Jörn Richter <joern.richter@th-koeln.de>",
]
readme = "README.md"
repository = "https://github.com/innovation-hub-bergisches-rheinland/showcase-3d-printer-anomaly-detection"

[tool.poetry.dependencies]
python = "3.8.6"
appdirs = "^1.4.4"
avro-python3 = "^1.10.2"
kafka-python = "^2.0.2"
PyYAML = "^5.4.1"
pandas = "^1.3.0"
scikit-learn = "0.23.1"
importlib-resources = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
bandit = "^1.7.0"
black = "^21.7b0"
bump2version = "^1.0.1"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
